Move utxo pruner interface package location from stores/pruner to stores/utxo/pruner
