+++ stores/utxo/pruner/interfaces.go
