+++ stores/utxo/sql/pruner/pruner_service_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/bsv-blockchain/teranode/settings"
-	"github.com/bsv-blockchain/teranode/stores/pruner"
+	"github.com/bsv-blockchain/teranode/stores/utxo/pruner"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
